# next-js-app/docs/swagger.yaml
openapi: "3.0.3"
info:
  version: "1.0.0"
  title: "Begehungstool API"
  description: >
    This API handles authentication, administration, and management of categories, subcategories, questions, departments, person types, and more for the Begehungstool (hygiene inspection tool).

  contact:
    name: "Support"
    email: "begehung@yourdomain.example"
servers:
  - url: "https://your-domain.example"  # Adjust to your actual domain
    description: "Production Server"
...
paths:
...
  /api/questions:
    get:
      tags:
        - "Question"
      summary: "Fetch questions with optional filtering."
      description: >
        Returns question data, including subcategory/category details.  
        Accepts optional search by text, department filtering, ID filtering, or `exclude` IDs.  
        Sorts results by category.priority then subcategory.priority.
      parameters:
        - name: department
          in: query
          schema:
            type: string
          required: false
          description: "Department ID to return only connected questions."
        - name: exclude
          in: query
          schema:
            type: string
          required: false
          description: "Semicolon-separated list of question IDs to exclude."
        - name: search
          in: query
          schema:
            type: string
          required: false
          description: "Search string to match `question` text (case-insensitive)."
        - name: id
          in: query
          schema:
            type: string
          required: false
          description: "Fetch only a single question by ID."
      responses:
        '200':
          description: "JSON array of matching questions or a single question if `id` is used."
        '500':
          description: "Server error or invalid query."

    post:
      tags:
        - "Question"
      summary: "Create a new question."
      description: >
        Admin only.  
        Body must have `{ question, critical, subcategory, type }`.  
        `type` can be `'Beobachtung' | 'Frage Personal' | 'Frage ärztliches Personal' | 'nicht anwendbar'`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                critical:
                  type: boolean
                subcategory:
                  type: number
                type:
                  type: string
              required: [question, subcategory]
      responses:
        '201':
          description: "Question created."
        '400':
          description: "Missing fields."
        '401':
          description: "Unauthorized."
        '500':
          description: "Database error."

    delete:
      tags:
        - "Question"
      summary: "Delete a question."
      description: >
        Admin only.  
        If question is connected to any department, fails with 400.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: "Question ID to delete."
      responses:
        '200':
          description: "Deleted successfully."
        '400':
          description: "Missing param or question still connected."
        '401':
          description: "Unauthorized."
        '500':
          description: "Database error."

    put:
      tags:
        - "Question"
      summary: "Update a question."
      description: >
        Admin only.  
        Body must have `{ id, question, critical, subcategory, type }`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: number
                question:
                  type: string
                critical:
                  type: boolean
                subcategory:
                  type: number
                type:
                  type: string
              required: [id, question, subcategory]
      responses:
        '200':
          description: "Successfully updated."
        '400':
          description: "Missing or invalid parameters."
        '401':
          description: "Unauthorized."
        '500':
          description: "Database error."
...
components:
  securitySchemes:
    CookieAuth:
      type: http
      scheme: bearer
      description: >
        This API mostly uses Supabase JWT tokens or server sessions.  
        Some admin routes require checking the user is in the `admin_users` table.
security: []